+++ Sources/ProjectSpec/ProductSubtype.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public enum ProductSubtype: String, Hashable, CaseIterable {
+    case widgetKitExtension = "widgetkit-extension"
+}